refactor(inventory): break out inventory types
Breaks out inventory types out of the inventory index.js.
